improve pretty print of if-else-elif
Summary:
Python has an interesting way to represent `elif` instances. ["elif clauses don’t have a special representation in the AST, but rather appear as extra If nodes within the orelse section of the previous one."](https://docs.python.org/3/library/ast.html)

Hence if an `orelse` field of an `If` node contains a single `if` statement then the parent `else` of the `If` node gets turned into an `elif` block with the contents of the `if`. However, if there is more than one item in the `orelse` field block (even if the first item is an `if`) then this logic does not apply and the `else` of the `If` node stays as an `else`.

```
if cond1 == 2:
    pass
else: # this is turned into an elif
    if cond2 == "py2":
        pass
    # as the block just contains an if

if cond1 == 2:
    pass
else: # this is not turned into elif
    if cond2 == "py2":
        pass
    x = 3 # because the block contains other things besides an if
```

Reviewed By: stroxler

Differential Revision:
D42668589

Privacy Context Container: L1152058

fbshipit-source-id: 3b1fcec330bca8411e4a29cd6fc705400316860b