pattern lists on lhs of assignment
Summary:
"Pattern lists" may occur on the left hand side of assignments. These consist of a comma seperated list of patterns and are represented in the AST as a tuple. Previously we were incorrectly transforming the lhs "pattern list" as a tuple (rather than just returning a tuple).

Here are some examples of "pattern list"s:
```
a, z = d
a, (b, c) = d
```

Reviewed By: stroxler

Differential Revision:
D42774256

Privacy Context Container: L1152058

fbshipit-source-id: e0acb4f64c8a593b94d926ee56800c5dad01c402