vendor in dependencies from hphp
Summary: Here we vendor in the dependencies ERRPY has on the Rust<->OCaml FFI provided by the Hack team under `/hphp/`. Note that we adjust the autocargo config so as to permit the automatic generation of the `Cargo.toml` files. Both the buck2 and cargo build now work ok

Reviewed By: grievejia

Differential Revision:
D42970729

Privacy Context Container: L1152058

fbshipit-source-id: dab0b8b9e0348c25b277725d994163f1709f8994