Add dune config
Summary:
Here we add the dune-project and dune files needed to build ERRPY both via dune internally and in the OSS build. Note that the dune build will invoke cargo to build the Rust code.

The cargo files are automatically generated by autocargo. So going forward we should only really need to keep the buck2 config updated and occationally tweak the dune config for OCaml changes.

We also update the `README.md` with the build instructions for the OSS release.

Reviewed By: grievejia

Differential Revision:
D42973110

Privacy Context Container: L1152058

fbshipit-source-id: b5c10ced7f5cf9ec6867453fe4a38cae9d3a2d74