report error nodes as syntax errors
Summary:
Here we report all `ERROR` nodes in the Tree-sitter CST as syntax errors. This enables Pyre to pass these errors on to an IDE for error highlighting whilst a valid, error-recovered AST is returned.

We also add some code in the recording function (`record_recoverable_error`) to ensure that `ERROR` nodes are recorded as `SyntaxError`'s. This code will be removed in a later diff where we filter out ERROR nodes from the CST to be walked (as they have already been marked as SyntaxError's).

Reviewed By: stroxler

Differential Revision: D42984426

fbshipit-source-id: ae7420a59482c5b86908966c7d802fcee0aa0db4