prevent keywords being useable as identifiers
Summary:
Here we prevent keywords being useable as identifiers. This probably should be a fix in tree-sitter but we can workaround this for now here.

In order to test this we define a new test suite: `test_invalid_syntax.py` to test where failure is expected.

Reviewed By: grievejia

Differential Revision:
D43333830

Privacy Context Container: L1152058

fbshipit-source-id: 031e76789b65d9d73be717cacf76fdf8313fcdb6