kw_defaults can be None - fix ast.ml
Summary:
As par the [AST spec](https://docs.python.org/3/library/ast.html#ast.arguments) - `kw_deaults` can be None:

> kw_defaults is a list of default values for keyword-only arguments. If one is None, the corresponding argument is required.

This was reflected in the `ast.rs` but not in the `ast.ml`. Which would lead to a crash of ERRPY when trying to marshal an AST across the FFI boundary between Rust and OCaml. 😵‍💫

Added a unit test for this. Generally speaking ERRPY may benefit from more rigours testing of the translation between the Rust and OCaml versions of the AST's.

Reviewed By: grievejia

Differential Revision:
D43429129

Privacy Context Container: L1152058

fbshipit-source-id: 222afadcc30abdcba0a9a9a117dbc0c5b07c2c27