unit test perf regression fix
Summary: Turns out that the new `ast_utils.format_ast_with_indentation` function for indenting AST's for easy reading and diffing is very slow. In this diff we differ calling of this function until we need a diffable or human readable output. We also make use of a list in order to hold the intermediate state of the indented AST and don't rely on string concatination.

Reviewed By: grievejia

Differential Revision:
D43682186

Privacy Context Container: L1152058

fbshipit-source-id: 06a57da3de22f5bb228462d16a62ab32db7a79f3