Add ByteStr Constant Type
Summary:
Presently ERRPY will output a `Str` for both byte and non-byte strings.  The consuming client (i.e. Pyre) then has to figure out if it's a byte or normal string based on the perfix of the String. This is not a very elegant way of doing things so here we introduce a new `ByteStr` type.

Note that we are still persisting the prefix (e.g. `b`) on the strings and we are not correctly handleing the different variants of prefix listed [here](https://docs.python.org/3/reference/lexical_analysis.html#strings) - i.e. `"b" | "B" | "br" | "Br" | "bR" | "BR" | "rb" | "rB" | "Rb" | "RB"`

This and other string related bugs will be resolved in later diffs. This is just the foundation.

Reviewed By: grievejia

Differential Revision: D44664849

fbshipit-source-id: 4efdc9efb0115aa3639dc8334cbdf8e806edfff1