Add string prefix categorization util
Summary:
Here we add a nice utility function `categorize_string` which is used in order to centralise and correctly implement the logic for determining if a string is a raw, byte, format or unicode string from its prefix or an appropriate combination thereof.

Note that we still have a way to go with the cleanup of the string implementation in ERRPY but this is a decent step in the path.

Reviewed By: stroxler

Differential Revision: D44682689

fbshipit-source-id: d557db280b5d141c3be1e5d8a5f8859eb45611e3