update grammar
Summary:
Here we do a few things:
 1. We update the parser grammar to the latest version in the tree-sitter-python github repo.
 2. Regenerate the parser. - `tree-sitter generate`
 3. Pass the tree-sitter-python unit tests.
 4. Make changes to `cast_to_ast.rs` to accommodate the update.
 5. Document how to run the tree-sitter-python generator and tests (`README.md`).

`cast_to_ast.rs` and `sitter.rs` is the only file needing close examination.

# Why are we doing this?
We will need to raise a PR on the tree-sitter-python github repo in order to upstream our changes to accommodate the match statement. But first we need to upgrade our version of `tree-sitter-python` to match the master branch of the github repo. This upgrade is needed because the master branch does not (yet) have an official release associated with it and so `ERRPY` (and the other project which depends on `tree-sitter-python`) at Meta have effectively fallen behind (through no fault of our own).

The changes to the grammar are found in `vendor/tree-sitter-python/grammar.js` - the other code in the vendor directory is autogenerated by `$ tree-sitter generate`.

Reviewed By: pradeep90

Differential Revision: D45069507

fbshipit-source-id: 0c2f7ebaed03fa2ef3393a8cfb97a14eba414e52