refactor expect tests code
Summary:
Here we refactor the `compare_recovered_ast_many` method to use a separate `check_vs_expects_file` method which wraps up the logic for comparing the output of a test vs an expected output committed to disk.

Three things can happen with expect tests:
 - If there is no expected results already then this is created using the obtained results and the test fails.
 - If the two test outputs differ (obtained and expected) then a new `expected_new` file is committed to disk and the test fails.
 - If the two test outputs match then this part of the test passes.

We do this refactoring now to make adding new tests which use this functionality in the diff stack cleaner to implement.

Note that we also do some refactoring of `check_vs_expects_file` to remove some printing/diffing logic on `pprint_ast` results which was not correctly implemented previously.

Reviewed By: grievejia

Differential Revision: D45302455

fbshipit-source-id: 2acf5a7e575f53753fee148821a0ce2d71cd449b