- MatchValue and Wildcards
Summary:
Here we add basic support for the `match` statement by permitting acceptance of the following sorts of input:

```
match x:
    case "some constant": # a constant value
        thing1()
    case 12: # another constant value
        thing2()
    case _: # a wildcard
        thing3()
```

In short, constant value patterns map to `MatchValue` and wildcards, `_` map to `MatchAs` with no arguments 🙄 (aside: seems like an odd design decision, why not have a `MatchWildcard`...?)

We're able to tick off a couple of the test cases, though the output for `unimplemented_syntax.pytest_expected_result` get s bit polluted, but this is resolved in later diffs in the stack.

Other cases are stubbed out for now.

Reviewed By: stroxler

Differential Revision: D45552208

fbshipit-source-id: 1d291d60272b30eec1de23b58d6181af039a70fd