match negative numbers
Summary:
Here we add support for negative numbers (integer and float) in pattern cases. e.g.:

```
match xx:
    case 3 | -3:
        pass
    case 3.0 | -3.0:
        pass
```

Added a new test case for this instance and refactored out the `case_literal_pattern` method from the body of `case_closed_pattern` - which was also missing a proper documentation string.

Note here that we had to tweak the grammar specification a bit, so there is some derived code in the diff.

Files to review:
 - `grammar.js`
 - `cst_to_ast.rs`
 - `ast_pattern_matching.pytest`
 - `pattern_matching.txt`

Other code is auto-generated from `tree-sitter generate` command.

Reviewed By: stroxler

Differential Revision: D45823848

fbshipit-source-id: 581c2953591bfd887454d710533c836392f8bcd4