- dotted name/identifiers
Summary:
Here we add `match` case pattern support for identifiers expressed as a single variable or in a dot delimited form: (`a.b.c`).

```
match x:
    case x:
        f=99
    case a.b.c.d:
        f=99
```

In the case of a single identifier, this is translated to a identifier wrapped into a `MatchAs`. Where there is more than one is translated into `Attribute` access pattern wrapped in a `MatchValue`.

We add a new more specific test case for this and can pass one of the existing ones.

Reviewed By: stroxler

Differential Revision: D45594374

fbshipit-source-id: 5138f8ab9f78ae1b890890c3cf7628c8e8e99cfe