case_group_pattern
Summary:
Here we add support for the case_group_pattern grammar rule in the tree-sitter-python grammar. This permits nesting of case patterns within `()`'s.. e.g.

```
match x:
    case [x] | (y):
        orthing()
    case [x] | (y) | ( a | b | c):
        # prety printer needs to add extra params for above
        orthing()
```

The existing case pattern matching code had to be refactored in places to directly return a `Pattern` object instead of a `PatternDesc` object because `case_group_pattern` needs to return a `Pattern` node which has line and column information relating to the nested pattern.

Some tweaking of the `format_vec_pattern` formatting function was required so as to avoid code duplication.

Added new test case for this specifically and existing tests are now passing

Reviewed By: stroxler

Differential Revision: D45647124

fbshipit-source-id: af05a5524a57c0fe23bee177fafba1c7bdf80dd2