- match support to tree-sitter-python grammar
Summary:
Here we add changes to the Python grammar definition used by tree-sitter within ERRPY. Note that Tree sitter is a parser generator so most of the code here is generated.

Previously the grammar permitted any `expression` to be used in place of a case pattern. This is problematic for two reasons:
 1. It permitted *any* expression to be used as case pattern which is not valid.
 2. For valid patterns it was difficult to transform them from a general `expression` in the CST into the form which we need them to be in (an AST) within ERRPY.

Besides the changes to `pattern_matching.txt`, all code here is generated based off changes to `grammar.js`, from running the `tree-sitter generate` command. The key files to review here are:
 - `grammar.js` - This contains the changes to support the match syntax.
 - `pattern_matching.txt` - Here we can see a richer CST output from tree-sitter-python given the changes to the grammar. We also add a few new test cases to improve thoroughness.

Once internally reviewed and landed I will start the process to upstream the change by raising a PR with the changes on the `tree-sitter-python` repo. There may be some small tweaks to the definition which we will have to make and we can incorporate these as they are raised (whilst making sure they work for ERRPY).

(see D45047065 for context)

Reviewed By: stroxler

Differential Revision: D45428697

fbshipit-source-id: fefed722d01941d5808edbce2b2bd0d81ff225e3