MatchClass
Summary:
Here we add support for class pattern matching:
```
match x:
    case Something(): # no args
        foo1()
    case Something.Else(): # more complex class name
        foo2()
    case Point2D(0, 1, 2): # three args
        foo3()
    case Point3D(x=0, y=0, z=0): # kw args
        foo4()
    case Point3D(34, x=0, y=0, z=0): # positional + kw args
        foo5()
    case Point2D(0, 1, 2,): # three args + trail comma
        foo6()
    case Point3D(x=0, y=0, z=0,): # kw args + trail comma
        foo7()
    case Point3D(34, x=0, y=0, z=0,): # positional + kw args + trail comma
        foo8()
```

This is the final piece of functionality which enables us to largely close out all of the missing functionality of the match support in ERRPY. This allows us to close out a lot of unit tests.

Tests wise, I've added a more detailed MatchClass test, tidied up `quick intro: literals 2`, and all the basic match tests now pass.

There are a few bugs left to fix which will be added in later diffs (mostly manifest via the unexpected failures in the `more_match` tests).

We remove the `unimplemented_syntax` test now because the `match` statement syntax is now supported.

Reviewed By: stroxler

Differential Revision: D45774084

fbshipit-source-id: a7c405108fcf583c66567c347c7e6f4a165d002b