remove colun offset bug workaround
Summary: In our unit tests, we compare the output AST of ERRPY with that of CPython. In earlier versions of CPython 3.10 there was a bug concerning column offsets for multiline f-strings. This has now been resolved so we can update ERRPY to remove the workaround introduced to resolve this bug as well.

Reviewed By: grievejia

Differential Revision: D46366088

fbshipit-source-id: 3175ec098837fa42c5c1880db0637cd218d899d1