Fix dune build
Summary: The OSS dune build was failing as the `Cargo.toml` was not generated correctly to support compilation of the C dependencies of the temporarily vendored in implementation of `tree-sitter-python`. This tweak to the `TARGETS` config fixes this problem, autocargo generates the `Cargo.toml` and now `dune test` runs ok.

Differential Revision: D46380033

fbshipit-source-id: 43e1109acd5860b94febfebf8ec49099cbaec113