filter out comment and error nodes
Summary:
Tree-sitter can emit `comment` and `error` nodes anywhere in its output CST's. This means that in theory our code to translate this into an AST needs to handle the possibility of a comment or error node being present anywhere in the CST! This is not really practical to implement in `cst_to_ast.rs` so instead here we add `node_wrapper` to filter out `ERROR` and `comment` notes in a new `Node` struct which wraps around the Tree sitter `Node`.

Within the `node_wrapper` we make use of an arena based tree allocator within `FilteredCST` for ease of understanding and future maintenance.

Key files to review:
 - `node_wrapper.rs` - main body of changes.
 - `cst_to_ast.rs` - Some refactoring to accommodate arena based tree allocator, but no new logic.

Rest is just plumbing and minor changes to error recovery in test results.

Reviewed By: stroxler

Differential Revision: D43717059

fbshipit-source-id: a4b3ba64fd01dcdac252bc49c013c2c4a1176d32