list_splat can be an expression
Summary:
Add support for star before variable on right hand size of assignment operator. e.g.:

```
f = *d,
```

This diff adds a small grammar change to support the above syntax. As with most of the grammar changes, the code is largely autogenerated. Code of specific interest to review is as follows:
 - `grammar.js` - Tweak to grammar to support the above syntax.
 - `tree-sitter-python/test/corpus/expressions.txt` - Addition to unit tests to support syntax.
 - `cst_to_ast.rs` - Change to permit translation to AST.
 - `ast_expressions.pytest` - Unit test within ERRPY supporting syntax.

Once we've made some more progress with the pattern match statement that work and this change will be upstreamed to `tree-sitter-python`

Reviewed By: grievejia

Differential Revision: D46640908

fbshipit-source-id: 13320c1c087c6872517d6cb16e1ea35c858a6430