tweak precedence operator on case_maybe_sequence_pattern
Summary:
Previously we were using the wrong precedence operator for `case_maybe_sequence_pattern`. Here we change this to `right`. This allows us to accept code such as the following:
```
match a:
    case a1, *a2, a3:
        pass
```

With `prec.left` tree-sitter-python would attempt to match the above code to the following CST:

```
(module
  (match_statement
    (identifier)
    (case_clause
      (case_open_sequence_pattern
        (case_maybe_star_pattern
          (case_pattern
            (case_or_pattern
              (case_closed_pattern
                (dotted_name
                  (identifier))))))
        (case_maybe_sequence_pattern
          (case_maybe_star_pattern
            (case_star_pattern
              (identifier)))))
      (case_pattern
        (case_or_pattern
          (case_closed_pattern
            (dotted_name
              (identifier)))))
      (block
        (pass_statement)))))
```

The `case_maybe_sequence_pattern` is not capturing the `a3` as a `case_maybe_star_pattern` instead the `case_clause` is capturing it which isn't quite what we want.

But we want this to be matched as the following which is achieved by using `prec.right` for `case_maybe_sequence_pattern` to make it "greedily" consume the final `a3` as a `case_maybe_star_pattern`:
```
(module
  (match_statement
    (identifier)
    (case_clause
      (case_open_sequence_pattern
        (case_maybe_star_pattern
          (case_pattern
            (case_or_pattern
              (case_closed_pattern
                (dotted_name
                  (identifier))))))
        (case_maybe_sequence_pattern
          (case_maybe_star_pattern
            (case_star_pattern
              (identifier)))
          (case_maybe_star_pattern
            (case_pattern
              (case_or_pattern
                (case_closed_pattern
                  (dotted_name
                    (identifier))))))))
      (block
        (pass_statement)))))
```

Quite a bit of code generated here. Key changes to review are in:
 - `grammar.js` - actual change
 - `ast_pattern_matching.pytest` - added unit test
 - `pattern_matching.txt` - added unit test

Reviewed By: grievejia

Differential Revision: D47114008

fbshipit-source-id: e80179cdbe7482c3036e9f035005d36173f03ab1