Rename `test` directory to `tests` and add bash functions
Summary: This diff renames the test director to `tests` to make use of the builtin `#[cfg(test)]`macro for `tests` directory and compiles files in this directory only when we run cargo test. This is also to align with the rust lang book [test organization practices](https://doc.rust-lang.org/book/ch11-03-test-organization.html#the-tests-directory).

Reviewed By: stroxler

Differential Revision: D47406318

fbshipit-source-id: f982b994f956da7512e74701d39fb2847bfc0736