Update tree-sitter-python grammar to github master
Summary:
Here we update the grammar, tests, and `scanner.c` to match the latest version on the [tree-sitter-python github repo](https://github.com/tree-sitter/tree-sitter-python)

This is a prerequisite to us upstreaming the pattern matching changes.

Hopefully once the pattern matching changes are upstreamed and released we can unvendor this code and revert to using the offical release and no more diffs like this will be needed.

I considered breaking this up into smaller changes but realistically the changes here are quite small so I thought it best to just do this as one diff for the sake of atomicity.

Files of interest include:
 * `cst_to_ast.rs` - only place where we do any work - This consists of updating definitions to keep them in sync with the grammar and tweaks to support changes related to the `await` statement
 * `node_wrapper.rs` - updated to exclude `line_continuation`'s

Other files of interest include:
 * `scanner.c` - migrated from c++ to c - ¯\_(ツ)_/¯ - also affects the `TARGETS` file
 * `grammar.js` - gramar changes - only significant change is some tweaking of the `await` statement
 * `vendor/tree-sitter-python/test/*.txt` - OSS community has improved tests a bit, changes here are mostly around the way strings are parsed.

Other code autogenerated or not interesting

Reviewed By: stroxler

Differential Revision: D47502340

fbshipit-source-id: ee178a90db01dde98c1e91f734278ef60ca085db