Refactor existing parsing flow to conditionally run a post-process step before `cst_to_ast` conversion
Summary:
# Overview
In this stack of diffs we look to address trailing period error recovery. When faced with a trailing period syntax errors we want to construct an AST that preserves information about the object before the period (this is essential for powering autocomplete functionality—object attribute access)

Currently, when ERRPY parses a line with a trailing period, it omits the line from the resulting AST constructed because tree-sitter parses the entire expression on the line as an ERROR node, which gets deleted in our CST to AST conversion process. This is unacceptable for autocomplete as it prevents us from providing useful information to the user about the object before the period. This is a very common case when developers are deciding which function of a class, option of an enum, attribute of an object, etc. to access. The goal here is to preserve as much information as possible about the line, including the object before the period, which is essential for powering autocomplete functionality.

This stack of diff will be split into the following steps:
1. **Refactor existing parsing flow to conditionally run a post-process step before `cst_to_ast` conversion.**
2. Refactor isize to usize for RecoverableErrorLocation and related structs
3. Extend support for rust regex
4. Implement trailing period post-processing and e2e test the source file to CST step
5. Replace injected token with an empty value attribute node in `cst_to_ast` conversion
6. Prefix match on empty string attribute node in Pyre binary
 ---

Reviewed By: jasontatton

Differential Revision: D47439535

fbshipit-source-id: 283686b13b8e116bb1e63e418bb7fe5391e7c3ea