bump errpy to Python 3.12
Summary:
this is needed to support parsing except* (exception groups) with errpy, since `except*` is unparseable for previous versions of the cpython parser.

there were some minor changes to the AST:
- additional fields for type params on function and class defs
- AST printing changes for parentheses around tuples, whitespace for lambdas, single vs double quotes for f-strings

the biggest change is that https://peps.python.org/pep-0617/ and (maybe) https://peps.python.org/pep-0701/ affected how f-strings are parsed.

previously, all child nodes of the f-string were marked with the same location as the f-string node. now, the locations for each of the child nodes is different, which required some extra math to handle all the edge cases.

this logic is fairly convoluted, especially considering the interactions with the previously-existing logic to calculate offsets and handle unicode. if we were planning to maintain this it would be good to rewrite it, but if we're deprecating this soon it's probably not worth it.

Reviewed By: samwgoldman

Differential Revision: D57511121

fbshipit-source-id: ab5742b286c65edc67a0095507afa1cdfddce046