third-party/rust: make sure num_enum has default features in default universe
Summary:
D54934898 also added it to the no-std universe and turned off default
features. This made no difference in practice because num_enum already had
dependants in the default universe needing its default features, but it's
better to make this explicit.

(Ideally there'd be a mechanism to detect that dependencies are actually
requesting the features they depend on, but we don't have that.)

Also make sure the default universe features are sorted.

Reviewed By: dreiss

Differential Revision: D58143089

fbshipit-source-id: 4aa6b16f16e20deb78dc09f445118d587b3ac740